requirements.txt
drf-spectacular~=0.27.2
drf-spectacular-sidecar~=2025.2.1
djangorestframework~=3.15.2
django-filter~=24.3
Django~=4.2
dbf~=0.99.9
django-environ~=0.11.2
djangorestframework_simplejwt~=5.5.0
django-cors-headers~=4.7.0
weasyprint==52.5
mssql-django==1.5
cx_Oracle==8.3.0
asgiref==3.8.1
pyodbc
